had to convert mp3 files to wav files
This was a problem before already, but I forgot to previously do it
in this script. The files from the itps corpus are called wav files
but are actually mp3 files, so we have to convert them. I am converting
them with ffmpeg through the command line here, because it was easiest
this way

+++ scripts/data_prep/annotation_data_prep.py
@@ -1,19 +1,60 @@
 #!/usr/bin/env python3
-from fastai.data.all import untar_data, get_files
+import os
+import shutil
+import subprocess
+
 import pandas as pd
-import shutil, os
+from fastai.data.all import get_files, untar_data
+from tqdm import tqdm
 
-ANNOTATION_DATA_URL = (
-    "https://www.dropbox.com/s/92tbhcdlymk5s0w/annotation_data.tar.gz?dl=1"
+from helpers import make_tarfile
+
+
+def is_this_ok():
+    answer = input("Is this OK with you? [y/n]")
+    if answer == "y":
+        return True
+    else:
+        return False
+
+
+def convert_mp3_wav(inn, outn):
+    res = subprocess.run(
+        [
+            "ffmpeg",
+            "-i",
+            inn,
+            "-f",
+            "wav",
+            "-acodec",
+            "pcm_s16le",
+            "-ar",
+            "16000",
+            "-ac",
+            "1",
+            f"{outn}",
+            "-y",
+        ]
+    )
+    return res
+
+
+ANNOTATION_DATA_URL_ORIG = (
+    "https://www.dropbox.com/s/qzvrx0c3rrmxxl3/annotation_data_initial.tar.gz?dl=1"
 )
+OUT_PATH = "/home/jjs/Dropbox/Share to iTPS AI-Team/train_data_repository/annotation_data.tar.gz"
 
-def main():
-    p = untar_data(ANNOTATION_DATA_URL, force_download=True)
+
+def main(force_download=False):
+    p = untar_data(ANNOTATION_DATA_URL_ORIG, force_download=force_download)
     print("Finished download")
-    jp_path = p / "アノテーションデータ/20210511_アノテーションデータ_full/20210511_Japanese_Delivery_Fin"
-    en_path = p / "アノテーションデータ/20210511_アノテーションデータ_full/20210511_English_Delivery_Fin"
-    shutil.move(str(jp_path), p/"jp/")
-    shutil.move(str(en_path), p/"en/")
+    jp_path = p / "20210511_Japanese_Delivery_Fin"
+    if os.path.exists(jp_path):
+        shutil.move(str(jp_path), p / "jp/")
+
+    en_path = p / "20210511_English_Delivery_Fin"
+    if os.path.exists(en_path):
+        shutil.move(str(en_path), p / "en/")
     print("Prepared directories")
     dfs = []
     for lang in ["en", "jp"]:
@@ -36,21 +77,25 @@ def main():
             if r["test"]:
                 src = r["filename"]
                 langfolder = src.parent.parent
-                new_dir = langfolder / "test" / src.name
-                if not os.path.exists(new_dir.parent):
-                    os.mkdir(new_dir.parent)
-                if os.path.exists(new_dir) and not os.path.exists(src):
+                dest = langfolder / "test" / src.name
+                if not os.path.exists(dest.parent):
+                    os.mkdir(dest.parent)
+                if os.path.exists(dest):
+                    print("continuing")
                     continue
-                shutil.move(src, new_dir)
+                convert_mp3_wav(src, dest)
+
             if not r["test"]:
                 src = r["filename"]
                 langfolder = src.parent.parent
-                new_dir = langfolder / "train" / src.name
-                if not os.path.exists(new_dir.parent):
-                    os.mkdir(new_dir.parent)
-                if os.path.exists(new_dir) and not os.path.exists(src):
+                dest = langfolder / "train" / src.name
+                if not os.path.exists(dest.parent):
+                    os.mkdir(dest.parent)
+                if os.path.exists(dest):
+                    print("continuing")
                     continue
-                shutil.move(src, new_dir)
+                convert_mp3_wav(src, dest)
+
         dfs = dfs + [df]
         print(f"Finished moving {lang} files")
 
@@ -60,7 +105,28 @@ def main():
         lambda x: x["filename"].parent.parent / x["dataset"] / x["filename"].name,
         axis=1,
     )
-    df.to_csv(p / "annotation_data.csv")
 
-if __name__ == '__main__':
-    main()
+    ok = df["filename"].apply(os.path.exists).all()
+    if not ok:
+        print(
+            f"{(~df['filename'].apply(os.path.exists)).sum()} files could not be converted correctly"
+        )
+        ok = is_this_ok()
+        if ok:
+            working_files = df["filename"].apply(os.path.exists)
+            df = df[working_files].copy()
+    if ok:
+        for d in (p / "en").ls():
+            if not d.name in ["train", "test"]:
+                print(d)
+                shutil.rmtree(d)
+        for d in (p / "jp").ls():
+            if not d.name in ["train", "test"]:
+                print(d)
+                shutil.rmtree(d)
+        df.to_csv(p / "annotation_data.csv")
+        make_tarfile(OUT_PATH, p)
+
+
+if __name__ == "__main__":
+    main(force_download=True)
